// packages/data/src/patches/7.3a/generated/champions.ts
export const GENERATED_CHAMPIONS: Champion[] = [
  {
// ...
    "abilities": {
      "passive": {
        "id": "aatrox-passive",
        "name": "DEATHBRINGER STANCE",
        "maxRank": 1,
        "cooldown": null,
        "castTime": 0,
        "damage": [],
        "flags": {}
      },
      "q": {
        "id": "aatrox-q",
        "name": "THE DARKIN BLADE",
        "maxRank": 4,
        "cooldown": {
          "byRank": [
            12,
            10,
            8,
            6
          ]
        },
        "castTime": 0,
        "damage": [
          {
            "type": "physical",
            "base": {
              "byRank": [
                10,
                40,
                70,
                100
              ]
            },
            "ratios": [
              {
                "stat": "totalAd",
                "value": 0.75
              }
            ],
            "tags": []
          }
        ],
        "flags": {},
        "cost": null
      },
      "w": {
        "id": "aatrox-w",
        "name": "INFERNAL CHAINS",
        "maxRank": 4,
        "cooldown": {
          "byRank": [
            15,
            14,
            13,
            12
          ]
        },
        "castTime": 0,
        "damage": [
          {
            "type": "physical",
            "base": {
              "byRank": [
                25,
                40,
                55,
                70
              ]
            },
            "ratios": [
              {
                "stat": "totalAd",
                "value": 0.4
              }
            ],
            "tags": []
          }
        ],
        "flags": {},
        "cost": null
      },
      "e": {
        "id": "aatrox-e",
        "name": "UMBRAL DASH",
        "maxRank": 4,
        "cooldown": {
          "byRank": [
            8,
            7,
            6,
            5
          ]
        },
        "castTime": 0,
        "damage": [],
        "flags": {},
        "cost": null
      },
      "r": {
        "id": "aatrox-r",
        "name": "WORLD ENDER",
        "maxRank": 3,
        "cooldown": {
          "byRank": [
            75,
            65,
            55
          ]
        },
        "castTime": 0,
        "damage": [],
        "flags": {},
        "cost": null
      }
    },
    "provenance": WRPOCKET_7_3A_PROVENANCE
  },
// ...
]